requirements.txt
streamlit
streamlit-chat
sqlalchemy
openai
python-dotenv